Test current SDI headers.
git-svn-id: https://svn.aros.org/svn/aros/trunk@53403 fb15a70f-31f2-0310-bbcc-cdcc74a49acc

+++ test/sdi/examples/hooks/example_hookuse.c
@@ -74,7 +74,7 @@ int main(void)
 
       ret = CallHook(&HelloHook, "portable Hook");
 
-      printf("%ld\n", (long int)ret);
+      printf("%ld\n", ret);
 
       DROPINTERFACE(IUtility);
     }
